Add files via upload
Filming in Virtual Production (VP) LED Volumes comes along with various undiscovered difficulties. To capture the right image in concern of color reproduction it is important to recognize the architecture of the LEDs. LED walls used nowadays are built of pixels, with each pixel consisting of three diodes which are producing red, green and blue light. This is based on the fact, that these LED walls are manufactured to be great WCG (Wide Color Gamut) displays rather than being optimized cinematic light sources. This results in an extremely narrow banded color spectrum with larger gaps in between the different visually possible wavelengths and a highly poor color rendition of the scene when LED walls are used as a light source. Non-neutral refelcting objects potentially have a higher color shift, if their color is not represented sufficiently in the emitted light spectrum. As it is always promised, that you can do image-based lighting with LED Volumes which nowadays still can’t reproduce a wide spectrum of colors, it should be considered as a main issue to analyse. In this thesis was investigated from which amount of light coming from the walls, the colors of the scene are being affected. The following chapters will analyse how lighting a scene with currently available LED walls effects color quality in-camera and to what amount cinematographers should use additional full spectral light sources to minimize the occuring color shifts. The examination leads to the fact, that as long as there are no LED walls with a more solid color spectrum for LED virtual production usecases, LED walls should not be used as a key-light in LED virtual production studios. To a certain extend the wall is a tolerable partial fill-light beside full spectrum lighting sources.